# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Mumu Bookkeeping API",
    "description": "Mumu 记账软件后端 API 接口"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "schemas": {
...
      "Book": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "book-123"
          },
          "name": {
            "type": "string",
            "example": "默认账本"
          },
          "balance": {
            "type": "number",
            "example": 12500.5
          },
          "isDefault": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "string",
                "format": "date-time"
              }
            ]
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "string",
                "format": "date-time"
              }
            ]
          }
        },
        "required": [
          "id",
          "name",
          "balance",
          "isDefault",
          "createdAt",
          "updatedAt"
        ]
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cat-123"
          },
          "name": {
            "type": "string",
            "example": "餐饮"
          },
          "type": {
            "type": "string",
            "enum": [
              "EXPENSE",
              "INCOME"
            ],
            "example": "EXPENSE"
          },
          "color": {
            "type": "string",
            "example": "#FFD8A8"
          },
          "bgColor": {
            "type": "string",
            "example": "#FFD8A8"
          },
          "textColor": {
            "type": "string",
            "example": "#7C2D12"
          },
          "createdAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "string",
                "format": "date-time"
              }
            ]
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "string",
                "format": "date-time"
              }
            ]
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "color",
          "bgColor",
          "textColor",
          "createdAt",
          "updatedAt"
        ]
      }
    },
...
  },
  "paths": {
...
    "/books": {
      "get": {
        "tags": [
          "Books"
        ],
        "summary": "获取所有账本列表",
        "responses": {
          "200": {
            "description": "账本列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Book"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Books"
        ],
        "summary": "创建新账本",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "旅游基金"
                  },
                  "balance": {
                    "type": "number",
                    "example": 0
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          }
        }
      }
    },
    "/books/{id}": {
      "put": {
        "tags": [
          "Books"
        ],
        "summary": "更新账本",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "修改后的账本"
                  },
                  "isDefault": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Books"
        ],
        "summary": "删除账本",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          }
        }
      }
    },
    "/categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "获取所有类别标签",
        "responses": {
          "200": {
            "description": "标签列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "创建分类标签",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "example": "奶茶"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "EXPENSE",
                      "INCOME"
                    ],
                    "example": "EXPENSE"
                  },
                  "color": {
                    "type": "string",
                    "example": "#FFD8A8"
                  },
                  "bgColor": {
                    "type": "string",
                    "example": "#FFD8A8"
                  },
                  "textColor": {
                    "type": "string",
                    "example": "#7C2D12"
                  }
                },
                "required": [
                  "name",
                  "type"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          }
        }
      }
    },
    "/categories/{id}": {
      "delete": {
        "tags": [
          "Categories"
        ],
        "summary": "删除分类标签",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          }
        }
      }
    },
...
  }
}